+++ test/formatters/retrieveResultFormatter.test.ts
@@ -5,6 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { join } from 'path';
 import * as sinon from 'sinon';
 import { expect } from 'chai';
 import { Logger } from '@salesforce/core';
@@ -13,6 +14,7 @@ import { cloneJson } from '@salesforce/kit';
 import { stubInterface } from '@salesforce/ts-sinon';
 import { getRetrieveResult } from '../commands/source/retrieveResponses';
 import { RetrieveCommandResult, RetrieveResultFormatter } from '../../src/formatters/retrieveResultFormatter';
+import { toArray } from '../../src/formatters/resultFormatter';
 
 describe('RetrieveResultFormatter', () => {
   const sandbox = sinon.createSandbox();
@@ -78,20 +80,21 @@ describe('RetrieveResultFormatter', () => {
       expect(formatter.getJson()).to.deep.equal(expectedInProgressResults);
     });
 
-    it.skip('should return expected json for a success with packages', async () => {
+    it('should return expected json for a success with packages', async () => {
+      const testPkg = { name: 'testPkg', path: join('path', 'to', 'testPkg') };
       const expectedSuccessResults: RetrieveCommandResult = {
         inboundFiles: retrieveResultSuccess.getFileResponses(),
-        packages: [],
+        packages: [testPkg],
         warnings: [],
         response: cloneJson(retrieveResultSuccess.response),
       };
-      const formatter = new RetrieveResultFormatter(logger, ux, {}, retrieveResultSuccess);
+      const formatter = new RetrieveResultFormatter(logger, ux, { packages: [testPkg] }, retrieveResultSuccess);
       expect(formatter.getJson()).to.deep.equal(expectedSuccessResults);
     });
 
     it('should return expected json for a success with warnings', async () => {
       const warnMessages = retrieveResultWarnings.response.messages;
-      const warnings = Array.isArray(warnMessages) ? warnMessages : [warnMessages];
+      const warnings = toArray(warnMessages);
       const expectedSuccessResults: RetrieveCommandResult = {
         inboundFiles: retrieveResultWarnings.getFileResponses(),
         packages: [],
@@ -127,10 +130,11 @@ describe('RetrieveResultFormatter', () => {
         .and.contain(`${options.waitTime} minutes`);
     });
 
-    it.skip('should output as expected for a Failure', async () => {
-      const formatter = new RetrieveResultFormatter(logger, ux, {}, retrieveResultInProgress);
+    it('should output as expected for a Failure', async () => {
+      const formatter = new RetrieveResultFormatter(logger, ux, {}, retrieveResultFailure);
+      sandbox.stub(formatter, 'isSuccess').returns(false);
+
       formatter.display();
-      expect(styledHeaderStub.called).to.equal(true);
       expect(logStub.called).to.equal(true);
       expect(tableStub.called).to.equal(false);
       expect(logStub.firstCall.args[0]).to.contain('Retrieve Failed due to:');
@@ -143,9 +147,9 @@ describe('RetrieveResultFormatter', () => {
       expect(styledHeaderStub.calledTwice).to.equal(true);
       expect(logStub.called).to.equal(true);
       expect(tableStub.calledOnce).to.equal(true);
-      expect(styledHeaderStub.firstCall.args[0]).to.contain('Retrieved Source Warnings');
+      expect(styledHeaderStub.secondCall.args[0]).to.contain('Retrieved Source Warnings');
       const warnMessages = retrieveResultWarnings.response.messages;
-      const warnings = Array.isArray(warnMessages) ? warnMessages : [warnMessages];
+      const warnings = toArray(warnMessages);
       expect(tableStub.firstCall.args[0]).to.deep.equal(warnings);
     });